requirements.txt
pdfminer.six
python-docx
spacy
streamlit
pandas
